Fixing running yb-ctl from a yugabyte-db source directory, and expanding the test (#2)
Restoring the logic for finding the installation directory so that it works for a yugabyte-db source directory, where bin/yb-ctl is now a wrapper script around submodules/yugabyte-installation/bin/yb-ctl. Also adding logic for finding an existing installation even if --install-if-needed is not specified and we are not running bin/yb-ctl from an extracted tarball or a yugabyte-db source directory.

Also, if initdb fails, the initdb log is being printed to console.

Testing:

Adding a test for running yb-ctl from a submodules/yugabyte-installation/bin subdirectory of a directory that also has YugaByte DB installed in build/latest.
Adding pycodestyle checking (on Linux only).