Handle OSError from missing postgres process
During `destroy`, stopping tservers involves looking for
`postmaster.pid` files that contain the PIDs of postgres processes.  In
certain cases, the postgres process may no longer exist.  The desired
behavior is to continue stopping the rest of the tserver and postgres
processes rather than throw an error.  Implement this behavior.