yb-ctl: avoid listen_ip default overriding ip_start (#34)
`--ip_start` specifies the ip addresses to use for addresses like
`rpc_bind_address` and `pgsql_proxy_bind_address`.  For example,
`--ip_start 7` with `--rf 3` will create masters and tservers each with
addresses `127.0.0.7`, `127.0.0.8`, and `127.0.0.9`.

`--listen_ip` specifies the ip address to use for listen addresses like
`pgsql_proxy_bind_address`.

When `--listen_ip` is not given, it takes on a default value of
`0.0.0.0` (for rf 1).  This is a recent feature from commit
69e8b995ef5880ac84198df25cea3d379ef15075 ([#4566] yb-ctl listen_ip
should default to 0.0.0.0 for rf 1 (#32)).  The problem is when
`--ip_start` is given but `--listen_ip` is not given: it still sets the
default value for `ClusterOptions.listen_ip` and makes things like
`pgsql_proxy_bind_address` use `0.0.0.0`.  Restore the old behavior by
**not** setting `ClusterOptions.listen_ip` to `0.0.0.0` in that case.

Close: yugabyte/yugabyte-db#4648